created public controller w/ burnettk
